refactor: move MapPinFactory pakage
