+++ src/main/java/com/gdg/planpal/domain/calendar/api/CalendarApi.java
@@ -1,0 +1,4 @@
+package com.gdg.planpal.domain.calendar.api;
+
+public class CalendarApi {
+}
