+++ src/main/java/com/gdg/planpal/domain/calendar/application/CalendarService.java
@@ -1,0 +1,4 @@
+package com.gdg.planpal.domain.calendar.application;
+
+public class CalendarService {
+}
